+++ packages/creem-sdk/src/lib/config.ts
@@ -64,7 +64,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "v1",
-  sdkVersion: "1.6.0",
-  genVersion: "2.918.4",
-  userAgent: "speakeasy-sdk/typescript 1.6.0 2.918.4 v1 creem",
+  sdkVersion: "1.6.1",
+  genVersion: "2.931.0",
+  userAgent: "speakeasy-sdk/typescript 1.6.1 2.931.0 v1 creem",
 } as const;